feat: show only available trade types in guide
